requirements.txt
transformers>=4.26.1
torch
pyzmq
tqdm